+++ zerver/tests/test_create_video_call.py
@@ -1,4 +1,7 @@
+from unittest import mock
+
 import responses
+from django.http import HttpResponseRedirect
 
 from zerver.lib.test_classes import ZulipTestCase
 
@@ -157,3 +160,57 @@ class TestVideoCall(ZulipTestCase):
             content_type="application/json",
         )
         self.assert_json_success(response)
+
+    def test_create_bigbluebutton_link(self) -> None:
+        with mock.patch('zerver.views.video_calls.random.randint', return_value="1"), mock.patch(
+             'zerver.views.video_calls.random.SystemRandom.choice', return_value="A"):
+            response = self.client_get("/json/calls/bigbluebutton/create")
+            self.assert_json_success(response)
+            self.assertEqual(response.json()['url'],
+                             "/calls/bigbluebutton/join?meeting_id=%22zulip-1%22&password=%22AAAAAAAAAA%22"
+                             "&checksum=%22697939301a52d3a2f0b3c3338895c1a5ab528933%22")
+
+    @responses.activate
+    def test_join_bigbluebutton_redirect(self) -> None:
+        responses.add(responses.GET, "https://bbb.example.com/bigbluebutton/api/create?meetingID=zulip-1&moderatorPW=a&attendeePW=aa&checksum=check",
+                      "<response><returncode>SUCCESS</returncode><messageKey/></response>")
+        response = self.client_get("/calls/bigbluebutton/join?meeting_id=%22zulip-1%22&password=%22a%22&checksum=%22check%22")
+        self.assertEqual(response.status_code, 302)
+        self.assertEqual(isinstance(response, HttpResponseRedirect), True)
+        self.assertEqual(response.url, "https://bbb.example.com/bigbluebutton/api/join?meetingID=zulip-1&password=a"
+                                       "&fullName=King%20Hamlet&checksum=7ddbb4e7e5aa57cb8c58db12003f3b5b040ff530")
+
+    @responses.activate
+    def test_join_bigbluebutton_redirect_wrong_check(self) -> None:
+        responses.add(responses.GET,
+                      "https://bbb.example.com/bigbluebutton/api/create?meetingID=zulip-1&moderatorPW=a&attendeePW=aa&checksum=check",
+                      "<response><returncode>FAILED</returncode><messageKey>checksumError</messageKey>"
+                      "<message>You did not pass the checksum security check</message></response>")
+        response = self.client_get("/calls/bigbluebutton/join?meeting_id=%22zulip-1%22&password=%22a%22&checksum=%22check%22")
+        self.assert_json_error(response, "Error authenticating to the Big Blue Button server.")
+
+    @responses.activate
+    def test_join_bigbluebutton_redirect_server_error(self) -> None:
+        # Simulate bbb server error
+        responses.add(responses.GET,
+                      "https://bbb.example.com/bigbluebutton/api/create?meetingID=zulip-1&moderatorPW=a&attendeePW=aa&checksum=check", "", status=500)
+        response = self.client_get(
+            "/calls/bigbluebutton/join?meeting_id=%22zulip-1%22&password=%22a%22&checksum=%22check%22")
+        self.assert_json_error(response, "Error connecting to the Big Blue Button server.")
+
+    @responses.activate
+    def test_join_bigbluebutton_redirect_error_by_server(self) -> None:
+        # Simulate bbb server error
+        responses.add(responses.GET,
+                      "https://bbb.example.com/bigbluebutton/api/create?meetingID=zulip-1&moderatorPW=a&attendeePW=aa&checksum=check",
+                      "<response><returncode>FAILURE</returncode><messageKey>otherFailure</messageKey></response>")
+        response = self.client_get(
+            "/calls/bigbluebutton/join?meeting_id=%22zulip-1%22&password=%22a%22&checksum=%22check%22")
+        self.assert_json_error(response, "Big Blue Button server returned an unexpected error.")
+
+    def test_join_bigbluebutton_redirect_not_configured(self) -> None:
+        with self.settings(BIG_BLUE_BUTTON_SECRET=None,
+                           BIG_BLUE_BUTTON_URL=None):
+            response = self.client_get(
+                "/calls/bigbluebutton/join?meeting_id=%22zulip-1%22&password=%22a%22&checksum=%22check%22")
+            self.assert_json_error(response, "Big Blue Button is not configured.")